rename ab_testing to the more common toggles
