+++ .circleci/config.yml
@@ -224,7 +224,7 @@ jobs:
       - run:
           name: Build lambda ZIP
           command: |
-            zip -r edge_lambda_origin.zip ab_testing.js origin.js redirector.js redirects.js wiRedirector.js
+            zip -r edge_lambda_origin.zip toggler.js origin.js redirector.js redirects.js wiRedirector.js
             mkdir -p /tmp/artifacts
             cp edge_lambda_origin.zip /tmp/artifacts
             ls /tmp/artifacts
